+++ example/lib/main.dart
@@ -4,7 +4,6 @@ import 'dart:developer';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:quran_library/quran_library.dart';
-import 'package:quran_library/src/audio/audio.dart';
 
 Future<void> main() async {
   await WidgetsFlutterBinding.ensureInitialized();