+++ lib/src/audio/surah_audio/widgets/surah_back_drop_widget.dart
@@ -167,7 +167,9 @@ class SurahBackDropWidget extends StatelessWidget {
                   isDark: isDark,
                 ),
 
-                const SizedBox(height: 20.0),
+                const Spacer(),
+                SurahPlayLandscapeWidget(
+                    style: style, isDark: isDark, languageCode: languageCode),
               ],
             ),
           ),